Server/Python/requirements.txt
anyio==3.6.1
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
fastapi==0.79.0
h11==0.13.0
idna==3.3
mysql-connector-python==8.0.30
protobuf==3.20.1
pycodestyle==2.9.1
pydantic==1.9.1
pymongo==4.2.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.3.0
uvicorn==0.18.2
